Rework cli screen life
