+++ srcs/cli_v2/srcs/config.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+@dataclass
+class Config:
+    ws_url: str
+    matchmaking_url: str
+    verify_cert: bool
