+++ srcs/cli_v2/srcs/services/game_service.py
@@ -1,17 +1,15 @@
-# srcs/services/game_service.py
-
 import os
 import httpx
 from .auth_service import AuthService
 
-BASE_URL = os.getenv("BASE_URL", "https://app.127.0.0.1.nip.io:8443")
-USER_MGMT_URL = os.getenv("USER_MGMT_URL", "https://api.app.127.0.0.1.nip.io:8443")
 
-VERIFY_HTTP_CERTIFICATE = False
+
 
 class GameService:
-    def __init__(self, auth: AuthService):
+    def __init__(self, app , auth: AuthService):
+
         self.auth = auth
+        self.app = app
 
     def _get_headers(self):
         headers = {"Content-Type": "application/json"}
@@ -21,12 +19,12 @@ class GameService:
         return headers
 
     async def create_room(self, username: str | None = None):
-        async with httpx.AsyncClient(verify=VERIFY_HTTP_CERTIFICATE) as client:
+        async with httpx.AsyncClient(verify=self.app.config.verify_cert) as client:
             payload = {}
             if username:
                 payload["username"] = username
             response = await client.post(
-                f"{USER_MGMT_URL}/api/room/create/",
+                f"{self.app.config.matchmaking_url}/api/room/create/",
                 headers=self._get_headers(),
                 json=payload
             )
@@ -35,21 +33,21 @@ class GameService:
             return response.json()
 
     async def join_room(self, room_code: str, username: str | None = None):
-        async with httpx.AsyncClient(verify=VERIFY_HTTP_CERTIFICATE) as client:
+        async with httpx.AsyncClient(verify=self.app.config.verify_cert) as client:
             payload = {"room_code": room_code}
             if username:
                 payload["username"] = username
             response = await client.post(
-                f"{USER_MGMT_URL}/api/room/join/",
+                f"{self.app.config.matchmaking_url}/api/room/join/",
                 headers=self._get_headers(),
                 json=payload
             )
             return response.json()
 
     async def check_room(self, room_code: str):
-        async with httpx.AsyncClient(verify=VERIFY_HTTP_CERTIFICATE) as client:
+        async with httpx.AsyncClient(verify=self.app.config.verify_cert) as client:
             response = await client.get(
-                f"{USER_MGMT_URL}/api/room/check/{room_code}/",
+                f"{self.app.config.matchmaking_url}/api/room/check/{room_code}/",
                 headers=self._get_headers()
             )
             return response.json()
