+++ srcs/cli_v2/srcs/services/services_registry.py
@@ -2,6 +2,6 @@ from .auth_service import AuthService
 from .game_service import GameService
 
 class ServiceRegistry:
-    def __init__(self):
-        self.auth = AuthService()
-        self.game = GameService(self.auth)
+    def __init__(self, app):
+        self.auth = AuthService(app)
+        self.game = GameService(app, self.auth)
